Updated the db structure to include the LocalProcurementAgent, BatchRequired, LocalProcurementLeadTime etc

+++ src/main/java/cc/altius/FASP/model/rowMapper/ShipmentListResultSetExtractor.java
@@ -52,11 +52,11 @@ public class ShipmentListResultSetExtractor implements ResultSetExtractor<List<S
                                         new SimpleObject(rs.getInt("PRODUCT_CATEGORY_ID"), new LabelRowMapper("PRODUCT_CATEGORY_").mapRow(rs, 1))))
                 );
                 s.setExpectedDeliveryDate(rs.getDate("EXPECTED_DELIVERY_DATE"));
-                s.setSuggestedQty(rs.getDouble("SUGGESTED_QTY"));
+                s.setSuggestedQty(rs.getInt("SUGGESTED_QTY"));
                 s.setProcurementAgent(new SimpleCodeObject(rs.getInt("PROCUREMENT_AGENT_ID"), new LabelRowMapper("PROCUREMENT_AGENT_").mapRow(rs, 1), rs.getString("PROCUREMENT_AGENT_CODE")));
                 s.setProcurementUnit(new SimpleObject(rs.getInt("PROCUREMENT_UNIT_ID"), new LabelRowMapper("PROCUREMENT_UNIT_").mapRow(rs, 1)));
                 s.setSupplier(new SimpleObject(rs.getInt("SUPPLIER_ID"), new LabelRowMapper("SUPPLIER_").mapRow(rs, 1)));
-                s.setQuantity(rs.getDouble("QUANTITY"));
+                s.setQuantity(rs.getInt("QUANTITY"));
                 s.setRate(rs.getDouble("RATE"));
                 s.setProductCost(rs.getDouble("PRODUCT_COST"));
                 s.setShipmentMode(rs.getString("SHIPPING_MODE"));
